+++ packages/ozone-helper/ozone-type/ozone/utility/index.ts
@@ -1,1 +1,1 @@
-export * from './FieldsPermission'
+export * from './fieldsPermissionUtility'
