--- packages/ozone-typescript-client/src/ozoneClient/clientConfiguration.ts
@@ -1,9 +1,0 @@
-import { OzoneCredentials } from '../ozoneCredentials/ozoneCredentials'
-
-export interface ClientConfiguration {
-	ozoneURL: string
-	ozoneInstanceId?: string
-	ozoneCredentials?: OzoneCredentials
-	webSocketsURL?: string
-	defaultTimeout?: number
-}
