--- packages/ozone-typescript-client/src/ozoneCredentials/ozoneCredentials.ts
@@ -1,11 +1,0 @@
-import { httpclient } from 'typescript-http-client'
-
-export interface AuthInfo {
-	principalClass: string,
-	principalId: string,
-	sessionId: string,
-}
-
-export interface OzoneCredentials {
-	authenticate(ozoneURL: string): Promise<AuthInfo>
-}
